Fix escaping spaces and special backslash escapes
Prior to this, all things that needed to be escaped other than single
and double quotes were escaped with full `\u` syntax.

My original intent was for things that have backslash-escapes (e.g.
newlines) to use their simpler backslash escape instead.

Unfortunately, neither the code nor the tests expressed that intent
correctly.

This changes 'escape' to operate in that manner.